// web/src/types/source.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type Source = { id: number, name: string, url: string, lastPub: string, lastPoll: string | null, ttl: number | null, favorite: boolean, minDate: string | null, createdAt: string, updatedAt: string, };